Rewrite of EllipticalGausisan source to directly fill the convolution grid for increased efficiency.
Addition of Ring, EllipticalDisk, and EllipticalRing spatial models.

Addition of convert_spatial_map for xml_parser to convert all spatial
models to the SpatialMap spatial model (for gtlike output).

Addition of copy() function to copy of a SpatialModel.

Vectorization of a few of the mappings between log and regular space.

Improved print formatting.

Degrees/radians Bugfix in template fitting